requirements.txt
fastapi
uvicorn
openai-whisper
googletrans==4.0.0-rc1
faiss-cpu
pydantic
requests
python-multipart
